server/requirements.txt
astroid==2.2.5
colorama==0.4.1
Django==2.1.7
django-cors-headers==2.5.2
djangorestframework==3.9.2
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.2
opencv-contrib-python==4.0.0.21
Pillow==5.4.1
pylint==2.3.1
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
